npcm7xx: net: mpdify npcm eth driver
